refactor(OrderMenuVerifierTest): Test를 위한 Subclass들 하나의 클래스로 리펙토링
